+++ src/Record.ts
@@ -153,7 +153,7 @@ export class WritableRecord {
 
     const headers: Record<string, string> = {
       "Content-Type": options.contentType ?? "application/octet-stream",
-      "Content-Length": contentLength.toString(),
+      "x-reduct-content-length": contentLength.toString(),
     };
 
     for (const [key, value] of Object.entries(options.labels ?? {})) {